ml_service/requirements.txt
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
flask==3.1.0
flask-cors==5.0.1
python-dotenv==1.0.1
gunicorn==23.0.0